+++ templates/project-ts/src/Add.test.ts
@@ -1,5 +1,6 @@
 import { AccountUpdate, Field, Mina, PrivateKey, PublicKey } from 'o1js';
 import { Add } from './Add';
+import { AddZkProgram } from './AddZkProgram';
 
 /*
  * This file specifies how to test the `Add` example smart contract. It is safe to delete this file and replace
@@ -8,7 +9,7 @@ import { Add } from './Add';
  * See https://docs.minaprotocol.com/zkapps for more info.
  */
 
-const proofsEnabled = false;
+const proofsEnabled = true;
 
 describe('Add', () => {
   let deployerAccount: Mina.TestPublicKey,
@@ -20,7 +21,10 @@ describe('Add', () => {
     zkApp: Add;
 
   beforeAll(async () => {
-    if (proofsEnabled) await Add.compile();
+    await AddZkProgram.compile({ proofsEnabled });
+    if (proofsEnabled) {
+      await Add.compile();
+    }
   });
 
   beforeEach(async () => {
@@ -45,23 +49,29 @@ describe('Add', () => {
     await txn.sign([deployerKey, zkAppPrivateKey]).send();
   }
 
-  it('generates and deploys the `Add` smart contract', async () => {
+  it('initilaizes the  `AddZKprogram`', async () => {
     await localDeploy();
-    const num = zkApp.num.get();
-    expect(num).toEqual(Field(1));
+
+    const { proof } = await AddZkProgram.init(Field(1));
+
+    expect(proof.publicOutput).toEqual(Field(1));
   });
 
-  it('correctly updates the num state on the `Add` smart contract', async () => {
+  it('correctly settles `AddZKprogram` state on the `Add` smart contract', async () => {
     await localDeploy();
+    const initialState = zkApp.num.get();
 
-    // update transaction
+    const init = await AddZkProgram.init(initialState);
+    const update = await AddZkProgram.update(initialState, init.proof);
+
+    // settleState transaction
     const txn = await Mina.transaction(senderAccount, async () => {
-      await zkApp.update();
+      await zkApp.settleState(update.proof);
     });
     await txn.prove();
     await txn.sign([senderKey]).send();
 
     const updatedNum = zkApp.num.get();
-    expect(updatedNum).toEqual(Field(3));
+    expect(updatedNum).toEqual(Field(1));
   });
 });